# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HabeshaShield API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Admin - list all users
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
  /admin/users/{id}:
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Admin - update a user (ban, change role)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Admin - delete a user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/reports:
    get:
      operationId: listReports
      tags: [admin]
      summary: Admin - list reported content
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
  /admin/reports/{id}:
    patch:
      operationId: resolveReport
      tags: [admin]
      summary: Admin - resolve a report
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportResolution"
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/security-alerts:
    get:
      operationId: listSecurityAlerts
      tags: [admin]
      summary: Admin - list security alerts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Security alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SecurityAlert"
  /admin/audit-logs:
    get:
      operationId: listAuditLogs
      tags: [admin]
      summary: Admin - list audit logs
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogListResponse"
  /admin/analytics:
    get:
      operationId: getAdminAnalytics
      tags: [admin]
      summary: Admin - get platform analytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAnalytics"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # AUTH
    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        displayName:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # USER
    User:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        displayName:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, artist, admin]
        isVerified:
          type: boolean
        isBanned:
          type: boolean
        followersCount:
          type: integer
        followingCount:
          type: integer
        createdAt:
          type: string
...
    Track:
      type: object
      required: [id, title, artistId, artistName, genre, youtubeId, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        artistId:
          type: string
        artistName:
          type: string
        artistAvatar:
          type: ["string", "null"]
        genre:
          type: string
        youtubeId:
          type: string
        thumbnail:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        likesCount:
          type: integer
        commentsCount:
          type: integer
        isLiked:
          type: boolean
        isTrending:
          type: boolean
        createdAt:
          type: string
...
    Report:
      type: object
      required: [id, targetType, targetId, reporterId, reason, status, createdAt]
      properties:
        id:
          type: string
        targetType:
          type: string
          enum: [track, comment, user, discussion]
        targetId:
          type: string
        reporterId:
          type: string
        reporterName:
          type: string
        reason:
          type: string
        details:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, resolved, dismissed]
        resolvedBy:
          type: ["string", "null"]
        createdAt:
          type: string
    ReportResolution:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [resolve, dismiss]
        notes:
          type: string

    # NOTIFICATIONS
...
    SecurityAlert:
      type: object
      required: [id, type, severity, message, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [failed_login, rate_limit, suspicious_activity, account_locked, xss_attempt]
        severity:
          type: string
          enum: [low, medium, high, critical]
        message:
          type: string
        ipAddress:
          type: ["string", "null"]
        userId:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
        isResolved:
          type: boolean
        createdAt:
          type: string
    AuditLog:
      type: object
      required: [id, action, userId, createdAt]
      properties:
        id:
          type: string
        action:
          type: string
        userId:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
        ipAddress:
          type: ["string", "null"]
        details:
          type: ["string", "null"]
        createdAt:
          type: string
    AuditLogListResponse:
      type: object
      required: [logs, total, page, totalPages]
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer

    # DASHBOARD
...
    AdminUserListResponse:
      type: object
      required: [users, total, page, totalPages]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
    AdminUserUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [user, artist, admin]
        isBanned:
          type: boolean
        isVerified:
          type: boolean
    AdminAnalytics:
      type: object
      required: [totalUsers, totalArtists, totalTracks, totalDiscussions, newUsersThisWeek, activeUsersToday, genreBreakdown, topTracks]
      properties:
        totalUsers:
          type: integer
        totalArtists:
          type: integer
        totalTracks:
          type: integer
        totalDiscussions:
          type: integer
        newUsersThisWeek:
          type: integer
        activeUsersToday:
          type: integer
        pendingReports:
          type: integer
        unresolvedAlerts:
          type: integer
        genreBreakdown:
          type: array
          items:
            type: object
            required: [genre, count]
            properties:
              genre:
                type: string
              count:
                type: integer
        topTracks:
          type: array
          items:
            $ref: "#/components/schemas/Track"