# lib/api-zod/src/generated/types/adminAnalytics.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HabeshaShield API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminAnalyticsGenreBreakdownItem } from "./adminAnalyticsGenreBreakdownItem";
import type { Track } from "./track";

export interface AdminAnalytics {
  totalUsers: number;
  totalArtists: number;
  totalTracks: number;
  totalDiscussions: number;
  newUsersThisWeek: number;
  activeUsersToday: number;
  pendingReports?: number;
  unresolvedAlerts?: number;
  genreBreakdown: AdminAnalyticsGenreBreakdownItem[];
  topTracks: Track[];
}