# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HabeshaShield API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: string;
  username: string;
  email: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  bio?: string | null;
  role: UserRole;
  isVerified?: boolean;
  isBanned?: boolean;
  followersCount?: number;
  followingCount?: number;
  createdAt: string;
}